Step 9-Component configuartion
