+++ Approot/webapp/index.js
@@ -1,11 +1,13 @@
 sap.ui.define([
-    "sap/ui/core/mvc/XMLView"
-], (XMLView) => {
-    "use strict";
+   "sap/ui/core/ComponentContainer"
+], (ComponentContainer) => {
+   "use strict";
 
-    XMLView.create({
-        viewName: "ui5.walkthrough.view.App"
-    }).then((oView) => {
-        oView.placeAt("content");
-    });
+   new ComponentContainer({
+      name: "ui5.walkthrough",
+      settings: {
+         id: "walkthrough"
+      },
+      async: true
+   }).placeAt("content");
 });
