--- quickstart/webapp/manifest.json
@@ -1,6 +1,0 @@
-{
-  "_version": "1.58.0",
-  "sap.app": {
-    "id": "ui5.quickstart"
-  }
-}
